+++ cpp/src/lib/messages/cucumber/messages/test_run_hook_started.cpp
@@ -13,6 +13,7 @@ test_run_hook_started::to_string() const
     cucumber::messages::to_string(oss, "id=", id);
     cucumber::messages::to_string(oss, ", test_run_started_id=", test_run_started_id);
     cucumber::messages::to_string(oss, ", hook_id=", hook_id);
+    cucumber::messages::to_string(oss, ", worker_id=", worker_id);
     cucumber::messages::to_string(oss, ", timestamp=", timestamp);
 
     return oss.str();
@@ -24,6 +25,7 @@ test_run_hook_started::to_json(json& j) const
     cucumber::messages::to_json(j, camelize("id"), id);
     cucumber::messages::to_json(j, camelize("test_run_started_id"), test_run_started_id);
     cucumber::messages::to_json(j, camelize("hook_id"), hook_id);
+    cucumber::messages::to_json(j, camelize("worker_id"), worker_id);
     cucumber::messages::to_json(j, camelize("timestamp"), timestamp);
 }