+++ javascript/src/messages.ts
@@ -672,6 +672,8 @@ export class TestRunHookStarted {
 
   hookId: string = ''
 
+  workerId?: string
+
   @Type(() => Timestamp)
   timestamp: Timestamp = new Timestamp()
 }